minor fixes

+++ src/integrations/tanstack-db/utils.ts
@@ -2,6 +2,8 @@ import { edgeCollection, taskCollection } from "./collections";
 import type { EdgeModel } from "./schema";
 
 export const deleteTaskWithDependencies = (taskId: string, edges: EdgeModel[]) => {
-  edgeCollection.delete(edges.map((edge) => edge.id));
+  if (edges.length > 0) {
+    edgeCollection.delete(edges.map((edge) => edge.id));
+  }
   taskCollection.delete(taskId);
 };
